feat(#138): move all Directives into separate package
