+++ src/main/java/org/eolang/jeo/representation/directives/DirectivesData.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.eolang.jeo.representation.asm;
+package org.eolang.jeo.representation.directives;
 
 import java.nio.ByteBuffer;
 import java.nio.charset.StandardCharsets;
