+++ src/main/java/org/eolang/jeo/representation/directives/DirectivesMethodProperties.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.eolang.jeo.representation.asm;
+package org.eolang.jeo.representation.directives;
 
 import java.util.Iterator;
 import java.util.Optional;
